refactor: api 경로 수정
